+++ src/operators/tensor/implementations/tensor_fp64x64.cairo
@@ -129,7 +129,7 @@ impl FP64x64Tensor of TensorTrait<FP64x64> {
         math::less::less(self, other)
     }
 
-    fn less_equal(self: @Tensor<FP64x64>, other: @Tensor<FP64x64>) -> Tensor<usize> {
+    fn less_equal(self: @Tensor<FP64x64>, other: @Tensor<FP64x64>) -> Tensor<i32> {
         math::less_equal::less_equal(self, other)
     }
